store db 수정!
