+++ .gitignore
@@ -1,5 +1,8 @@
 .idea/
 venv/
+
+**/migrations/
+
 .DS_Store
 # Created by https://www.toptal.com/developers/gitignore/api/python,django
 # Edit at https://www.toptal.com/developers/gitignore?templates=python,django
